% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erfcinv.R
\name{erfcinv}
\alias{erfcinv}
\title{Inverse Error Function Complement}
\usage{
erfcinv(sigma)
}
\arguments{
\item{sigma}{The partition coefficient of an eluted peak}
}
\value{
The inverse error function complimentary
}
\description{
The inverse error function complement
[https://en.wikipedia.org/wiki/Error_function](https://en.wikipedia.org/wiki/Error_function)
}

