% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{partition_coef}
\alias{partition_coef}
\title{Partition Coefficient}
\usage{
partition_coef(vs = 65, vo = 8.21, vt = 21.54)
}
\arguments{
\item{vs}{The observed elution volume}

\item{vo}{The void volume}

\item{vt}{The total column volume}
}
\value{
sigma: The partition coefficient of the elucted peak/fraction
}
\description{
The partition coefficient (s) was calculated for each
standard protein from the observed elution volume (Vs), total column volume (Vt),
and void volume (Vo)
}
\examples{
## calculate the partition coefficient of an elucted peak on a column with total
column volume of 120 mL and void volume of 39 mL.

partition(vs =65, vo=39, vt=120)
}

