% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_std_data.R
\docType{data}
\name{protein_std}
\alias{protein_std}
\title{Standard proteins of known Stokes radius}
\format{A data frame with 5 rows and 3 variables:
\describe{
 \item{name}{protein standard name}
 \item{mw}{molecular weight}
 \item{ve}{elution volume on a gel filtration column of superdex 200}
 \item{radii}{hydrodynamic radius}
}}
\usage{
protein_std
}
\description{
A dataset containing the a protein standard purchase from biorad with a cat
#151-1901.
}
\keyword{datasets}

