% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R, R/radius-package.r
\docType{package}
\name{radius}
\alias{radius}
\alias{radius-package}
\title{stokes radius}
\usage{
radius(ve, x)
}
\description{
Stokes radius is also call hydrodynamic radius for polymer.

Determine the hydrodynamic radius by gel filtration chromatography
}
\details{
## Ref:

- 1. \code{\url{https://en.wikipedia.org/wiki/Hydrodynamic_radius}}

- 2. \code{\href{http://www.pnas.org/content/105/2/482.full?tab=ds}{Supporting Information}}
}

